+++ CHANGELOG.md
@@ -10,6 +10,7 @@ unreleased
 - Display an open dialog on launch if so set even when iCloud storage is disabled.
 - Improve stability on text encoding change.
 - Refine dialog messages on text encoding change.
+- Change sidebar behavior to close inward when it was opened inward because of insufficient space.
 - [trivial] Hide insertion point in shortcut input fields in Key Binding pane.
 - [trivial] Add `enablesAsynchronousSaving` hidden default key that enables asyncronous saving.